Add Auto Commit Feature
This is something that Github Desktop does that I think is neat and
useful.

If there's nothing insightful about the commit you are making and you
are committing one file, use `arbor commit` to automatically commit a
single file.

The commit message is the filename with an imperative prefix (e.g. Add
file.txt).